% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceNames.R
\docType{data}
\name{distanceNames}
\alias{distanceNames}
\title{A vector containing names of the calculated distance measures between fuzzy numbers.}
\format{
An object of class \code{character} of length 3.
}
\usage{
distanceNames
}
\value{
This function returns a vector of strings.
}
\description{
`distanceNames` is a vector containing names of the calculated distance measures between fuzzy numbers.
}
\examples{

# check the names

distanceNames

}
\keyword{datasets}
