% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorTypes.R
\docType{data}
\name{errorTypes}
\alias{errorTypes}
\title{A vector containing names of the types of the calculated errors.}
\format{
An object of class \code{character} of length 5.
}
\usage{
errorTypes
}
\value{
This function returns a vector of strings.
}
\description{
`errorTypes` is a vector containing names of the types of the calculated errors.
}
\examples{

# check the names

errorTypes

}
\keyword{datasets}
