% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuresTypes.R
\docType{data}
\name{measuresTypes}
\alias{measuresTypes}
\title{A vector containing names of the types of the calculated statistical measures.}
\format{
An object of class \code{character} of length 12.
}
\usage{
measuresTypes
}
\value{
This function returns a vector of strings.
}
\description{
`measuresTypes` is a vector containing names of the types of the calculated statistical measures.
}
\examples{

# check the names

measuresTypes

}
\keyword{datasets}
