% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodNames.R
\name{methodNames}
\alias{methodNames}
\title{A vector containing names of the resampling methods.}
\usage{
methodNames
}
\value{
This function returns a vector of strings.
}
\description{
`methodNames` is a vector containing names of the resampling methods.
}
\examples{

# check the names

methodNames

}
