% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsNames.R
\docType{data}
\name{setsNames}
\alias{setsNames}
\title{A vector containing names of the subsets used in the epistemic tests.}
\format{
An object of class \code{character} of length 3.
}
\usage{
setsNames
}
\value{
This function returns a vector of strings.
}
\description{
`setsNames` is a vector containing names of the subsets used in the epistemic tests.
}
\examples{

# check the names

setsNames

}
\keyword{datasets}
