% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.impTest.R
\name{summary.impTest}
\alias{summary.impTest}
\title{Print summary of the benchmark for the imputation method.}
\usage{
\method{summary}{impTest}(object, ...)
}
\arguments{
\item{object}{Object of S3 class \code{impTest}.}

\item{...}{Additional parameters passed to other functions.}
}
\description{
Print summary of the benchmark for the imputation method.
}
