% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.metComp.R
\name{summary.metComp}
\alias{summary.metComp}
\title{Print summary of the comparison of the imputation methods.}
\usage{
\method{summary}{metComp}(object, ...)
}
\arguments{
\item{object}{Object of S3 class \code{metComp}.}

\item{...}{Additional parameters passed to other functions.}
}
\description{
Print summary of the comparison of the imputation methods.
}
