% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsNames.R
\docType{data}
\name{testsNames}
\alias{testsNames}
\title{A vector containing names of the types of the applied epistemic tests.}
\format{
An object of class \code{character} of length 3.
}
\usage{
testsNames
}
\value{
This function returns a vector of strings.
}
\description{
`testsNames` is a vector containing names of the types of the applied epistemic tests.
}
\examples{

# check the names

testsNames

}
\keyword{datasets}
