% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodNames.R
\docType{data}
\name{methodNames}
\alias{methodNames}
\title{A vector containing names of the resampling methods.}
\format{
An object of class \code{character} of length 4.
}
\usage{
methodNames
}
\value{
This function returns a vector of strings.
}
\description{
`methodNames` is a vector containing names of the resampling methods.
}
\examples{

# check the names

methodNames

}
\keyword{datasets}
